// src/lit.smabu.react/src/types/domain/create-payment-command.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LIT.Smabu.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { CustomerId } from './customer-id';
import { InvoiceId } from './invoice-id';
import { PaymentDirection } from './payment-direction';
import { PaymentId } from './payment-id';
 /**
 * 
 *
 * @export
 * @interface CreatePaymentCommand
 */
export interface CreatePaymentCommand {

    /**
     * @type {PaymentId}
     * @memberof CreatePaymentCommand
     */
    paymentId: PaymentId;

    /**
     * @type {PaymentDirection}
     * @memberof CreatePaymentCommand
     */
    direction: PaymentDirection;

    /**
     * @type {Date}
     * @memberof CreatePaymentCommand
     */
    accountingDate: Date;

    /**
     * @type {string}
     * @memberof CreatePaymentCommand
     */
    details?: string | null;

    /**
     * @type {string}
     * @memberof CreatePaymentCommand
     */
    payer?: string | null;

    /**
     * @type {string}
     * @memberof CreatePaymentCommand
     */
    payee?: string | null;

    /**
     * @type {CustomerId}
     * @memberof CreatePaymentCommand
     */
    customerId?: CustomerId;

    /**
     * @type {InvoiceId}
     * @memberof CreatePaymentCommand
     */
    invoiceId?: InvoiceId;

    /**
     * @type {string}
     * @memberof CreatePaymentCommand
     */
    referenceNr: string | null;

    /**
     * @type {Date}
     * @memberof CreatePaymentCommand
     */
    referenceDate?: Date | null;

    /**
     * @type {number}
     * @memberof CreatePaymentCommand
     */
    amountDue?: number;

    /**
     * @type {Date}
     * @memberof CreatePaymentCommand
     */
    dueDate?: Date | null;

    /**
     * @type {boolean}
     * @memberof CreatePaymentCommand
     */
    markAsPaid?: boolean | null;
}